requirements.txt
albumentations
catalyst
efficientnet_pytorch
numpy
pandas
scipy
scikit-image
torch
tqdm
tensorboard
matplotlib
ipywidgets
flask
flask-session
python-dotenv
tensorflow
efficientnet
jpegio
seaborn==0.8.1